Extract Records into separate class
